put more instrumentation in malloc.c, recording all request and frees from the application. bug fixing. refine.sh can be used to get the request sequence, stored in curr/ dir.

+++ myga/clean.sh
@@ -1,10 +1,12 @@
-cd curr
-rm *
-cd ..
+if [ -d curr ]; then
+	cd curr
+	rm *
+	cd ..
+fi
 
 containsElement () {
   local e
-  for e in "run.sh" "runall.sh" "callgrind.sh" "clean.sh" "malloc.c"; do [[ "$e" == "$1" ]] && return 0; done
+  for e in "run.sh" "runall.sh" "callgrind.sh" "clean.sh" "malloc.c" "refine.sh" "refineall.sh"; do [[ "$e" == "$1" ]] && return 0; done
   return 1
 }
 
@@ -13,6 +15,7 @@ for f in `ls`; do
 		containsElement $f
 		if [ $? -eq 1 ]; then
 			rm $f
+			#echo $f
 		fi
 	fi
 done
